Basic support for editing multiple rules.
